iccsdfai2025/library/requirements.txt
Flask==2.0.3
PyJWT==2.4.0
cryptography==38.0.4
Werkzeug==2.0.3
requests